// pharmacy-ui/src/app/services/models/product.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HealthCondition } from '../models/health-condition';
import { ProductCategory } from '../models/product-category';
export interface Product {
  active?: boolean;
  category?: ProductCategory;
  conditions?: Array<HealthCondition>;
  createdDate?: string;
  description?: string;
  expiryDate?: string;
  id?: number;
  image?: string;
  lastModifiedDate?: string;
  manufacturer?: string;
  name?: string;
  price?: number;
  soldCount?: number;
  stockQuantity?: number;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
  viewCount?: number;
}